# Refactor the subquery code and fix outer condition queries