@jsternberg 2017-03-01
@rkuchan is it possible for you to test this out with a few scenarios? I added some tests, but some human eyes would also be good.

@jwilder I'm not 100% confident in the performance characteristics of this, but I give my rationale for why I think it's a positive change anyway in the PR. What do you think?