@jsternberg 2017-03-02
This isn't a high priority for testing. We're going to push it to the 1.3 release since it's more involved than other bugfixes.