@jsternberg 2017-03-07
I still need to expand the benchmarks to cover more aspects, but the initial numbers are not promising for the simplest case:

```
benchmark                                     old ns/op     new ns/op     delta
BenchmarkSelect_Subquery_Aggregate_1K-8       785914        832085        +5.87%
BenchmarkSelect_Subquery_Aggregate_100K-8     62048569      72915654      +17.51%

benchmark                                     old allocs     new allocs     delta
BenchmarkSelect_Subquery_Aggregate_1K-8       7070           7077           +0.10%
BenchmarkSelect_Subquery_Aggregate_100K-8     700076         700083         +0.00%

benchmark                                     old bytes     new bytes     delta
BenchmarkSelect_Subquery_Aggregate_1K-8       411989        412261        +0.07%
BenchmarkSelect_Subquery_Aggregate_100K-8     40804498      40804770      +0.00%
```

That may not be an acceptable speed loss. Especially with 100K points.