@jsternberg 2017-04-17
I am so confused, but rerunning the benchmarks that I have locally on my computer gives me this now:

```
jsternberg:[~/g/s/g/i/influxdb] $ (tmp $) benchcmp old.txt new.txt
benchmark                                     old ns/op     new ns/op     delta
BenchmarkSelect_Subquery_Aggregate_1K-8       794322        783705        -1.34%
BenchmarkSelect_Subquery_Aggregate_100K-8     63811348      61522037      -3.59%

benchmark                                     old allocs     new allocs     delta
BenchmarkSelect_Subquery_Aggregate_1K-8       7070           7077           +0.10%
BenchmarkSelect_Subquery_Aggregate_100K-8     700076         700083         +0.00%

benchmark                                     old bytes     new bytes     delta
BenchmarkSelect_Subquery_Aggregate_1K-8       411989        412261        +0.07%
BenchmarkSelect_Subquery_Aggregate_100K-8     40804498      40804770      +0.00%
```

If true, this is definitely acceptable. I'm also trying to use the same old technique for subqueries that I was using, but trying to get the changes I made in this query and I am finding it very difficult to keep track of the code.